refactor features and add hook for dataset features
