+++ phoebe/features/__init__.py
@@ -1,0 +1,2 @@
+from .dataset_features import *
+from .component_features import *
